Moved and improved get_sources function
