+++ holoviews/plotting/bokeh/element.py
@@ -23,11 +23,12 @@ from bokeh.plotting.helpers import _known_tools as known_tools
 from ...core import (Store, HoloMap, Overlay, DynamicMap,
                      CompositeOverlay, Element, Dimension)
 from ...core.options import abbreviated_exception, SkipRendering
+from ...core.spaces import get_sources
 from ...core import util
 from ...element import RGB
 from ...streams import Stream, RangeXY, RangeX, RangeY
 from ..plot import GenericElementPlot, GenericOverlayPlot
-from ..util import dynamic_update, get_sources, attach_streams
+from ..util import dynamic_update, attach_streams
 from .plot import BokehPlot, TOOLS
 from .util import (mpl_to_bokeh, convert_datetime, update_plot, get_tab_title,
                    bokeh_version, mplcmap_to_palette, py2js_tickformatter,
